refactor: Restructure Forget Password screen and update assets
